+++ src/base/Win.cpp
@@ -2364,7 +2364,7 @@ Str ReadIStream(IStream* stream) {
     }
 
     int n = (int)stat.cbSize.QuadPart;
-    char* d = AllocArray<char>((int)n + (int)sizeof(WCHAR));
+    char* d = AllocArray<char>(n + (int)sizeof(WCHAR));
     if (!d) {
         return {};
     }
